FiberUpdateView

+++ Graph/Commands/Fiber/UpdateFiber.cs
@@ -5,6 +5,6 @@ namespace Iit.Fibertest.Graph.Commands
     public class UpdateFiber
     {
         public Guid Id { get; set; }
-        public double UserInputedLength { get; set; }
+        public int UserInputedLength { get; set; }
     }
 }
